Add Colemak layer and automated language switch
I've moved my Colemak-DH layout out of being handled by OS into my keyboard firmware, which allows me to use the same layout on any PC I connect to. I've added a temporary switch back to QWERTY for anything that might need it. I've also added a keycode for switching languages: when doing so, besides triggering a universal language switch combination (Win+Space, works out of the box in Windows and is the default in many Linux distros), it switches to QWERTY when switching to the second language, so that it won't mess with that keyboard layout. When switching back to English, it switches to Colemak-DH. If anything happens and Colemak ends up being used with the non-English layout or QWERTY ends up while typing in English, I can switch between the two manually. The language switch will still work correctly after this, as it keeps track of what layout is default.